feat(input): 容量约定 v2——动作位 u16→u32，预留位同宽保留（8B/人）
- ActionInput = { buttons: u32, _pad: u32 }：动作位翻倍（7/32 在用），
  参数预留位保留且加宽到 u32，结构体无隐式填充、每个字节都有名字
- PlayerState.input 同步加宽 u32（译码目的地同宽，防位 16..32 静默截断）
- 宏/EDGE_MASK/编译期断言/测试全部随型；位号与词表指纹不变（词表≠容器）
- 布局测试钉死 v2：size_of ActionInput == 8、InputFrame == 4 + 8×MAX_PLAYERS
- InputFrame 尺寸变更 = 格式变更，落在 M3 回放格式出生前的免费窗口
- 金向量校验和流因 PlayerState.input 位宽变化而变（预期；闸门无 committed
  基线、仅跨平台互比），110 单测全绿含双跑确定性

+++ crates/stg-core/src/player.rs
@@ -44,7 +44,7 @@ pub struct PlayerState {
     pub facing: i8, // 纯表现，照样入校验和（P6）
     pub hit_radius: Fx,
     pub graze_radius: Fx,
-    pub input: u16,
+    pub input: u32,
     pub life_state: u8,
     pub state_timer: u16,
     pub invuln: u16,
